quote bin name for default Rust build commands
if you make a mistake and use a name in spaces in your Cargo.toml, *and* don't have a build command, the CLI will construct a string like `cargo build --bin ixi test service`, which parses as "build the binary ixi and run the command test". cargo can't make sense of it, and errors about the word `test`. it's a little weird confusing to debug.